+++ Tomato/Sources/LetterList/Letter.swift
@@ -1,0 +1,13 @@
+//
+//  Letter.swift
+//  Tomato
+//
+//  Created by Daye on 1/3/25.
+//
+
+import Foundation
+
+struct Letter: Identifiable {
+    var id = UUID()
+    var recipient: String
+}
